Rename exception package to common package
