+++ src/main/java/duke/Duke.java
@@ -1,7 +1,7 @@
 package duke;
 
 import duke.command.Command;
-import duke.exception.DukeException;
+import duke.common.DukeException;
 import duke.parser.Parser;
 import duke.storage.Storage;
 import duke.task.TaskList;
