+++ src/main/java/duke/command/AddCommand.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.exception.DukeException;
+import duke.common.DukeException;
 import duke.storage.Storage;
 import duke.task.Task;
 import duke.task.TaskList;
